requirements.txt
joblib==1.4.2
numpy==2.0.0
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-surprise==1.1.4
scipy==1.14.0
six==1.16.0
tzdata==2024.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
scikit-learn==1.3.1
